# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cohort Discovery Service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/custodians": {
            "get": {
                "tags": [
                    "Custodians"
                ],
                "summary": "Get all custodians",
                "operationId": "9059e89c8b4f8fbe189466be80f63471",
                "responses": {
                    "200": {
                        "description": "List of custodians",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Custodian"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Custodians"
                ],
                "summary": "Create a new custodian",
                "operationId": "db138a1bb1458a88f3b521e42deec775",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Custodian"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Custodian created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Custodian"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/custodians/{id}": {
            "get": {
                "tags": [
                    "Custodians"
                ],
                "summary": "Get a custodian by ID",
                "operationId": "e3b0323c4a147591ee7bf1df6f7d0012",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custodian found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Custodian"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Custodian not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Custodians"
                ],
                "summary": "Update a custodian",
                "operationId": "96ef7c8ef5e377142ec5bba4df2481b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Custodian"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Custodian updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Custodian"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Custodian not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Custodians"
                ],
                "summary": "Delete a custodian",
                "operationId": "42a0fb2fc35fa3cd4737634e94540ac9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custodian deleted"
                    },
                    "404": {
                        "description": "Custodian not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Custodian": {
                "title": "Custodian",
                "required": [
                    "name",
                    "street_address",
                    "city",
                    "postal_code",
                    "country"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Health Data Custodian"
                    },
                    "street_address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "city": {
                        "type": "string",
                        "example": "London"
                    },
                    "postal_code": {
                        "type": "string",
                        "example": "SW1A 1AA"
                    },
                    "country": {
                        "type": "string",
                        "example": "United Kingdom"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://custodian.org"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@custodian.org"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+44 1234 567890"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-06T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-06T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}